+++ vpn/src/service/tunnel/client_tunnel/quic_tunnel.rs
@@ -1,5 +1,6 @@
 use std::{sync::Arc, time::Duration};
 
+use anyhow::Result;
 use futures::{channel::mpsc::Sender, Stream};
 use quinn::{crypto::rustls::QuicClientConfig, ClientConfig, Endpoint};
 use rustls::pki_types::{ServerName, UnixTime};
@@ -8,8 +9,8 @@ use tonic::transport::CertificateDer;
 use tracing::{error, info, trace};
 
 use crate::{
-    interval, util::channel_bus, ClientMsg, DataCrypt, QuicClientStreamGenerator, Tunnel, VpnError,
-    HEARTBEAT_INTERVAL_MS,
+    client::run_client, interval, util::channel_bus, ClientMsg, ClientQuicConn, DataCrypt, Tunnel,
+    VpnError, HEARTBEAT_INTERVAL_MS,
 };
 
 pub struct QuicTunnel;
@@ -26,9 +27,9 @@ impl QuicTunnel {
             loop {
                 match quic_tunnel_core_task(
                     server_addr.clone(),
+                    crypt.clone(),
                     &mut msg_stream,
                     main_sender_clone.clone(),
-                    crypt.clone(),
                 )
                 .await
                 {
@@ -46,13 +47,7 @@ impl QuicTunnel {
     }
 }
 
-async fn quic_tunnel_core_task<S: Stream<Item = ClientMsg> + Unpin>(
-    server_addr: String,
-    msg_stream: &mut S,
-    main_sender_tx: Sender<ClientMsg>,
-    crypt: Arc<Box<dyn DataCrypt>>,
-) -> Result<(), VpnError> {
-    trace!("Tcp tunnel core task start");
+async fn get_quic_stream(server_addr: String) -> Result<ClientQuicConn, VpnError> {
     let mut endpoint = Endpoint::client("0.0.0.0:0".parse().unwrap())?;
 
     endpoint.set_default_client_config(ClientConfig::new(Arc::new(QuicClientConfig::try_from(
@@ -68,37 +63,24 @@ async fn quic_tunnel_core_task<S: Stream<Item = ClientMsg> + Unpin>(
         .unwrap()
         .await
         .unwrap();
-    println!("[client] connected: addr={}", connection.remote_address());
+    info!("[client] connected: addr={}", connection.remote_address());
 
-    // Split client to Server stream
-    let (mut read_stream, mut write_stream) =
-        QuicClientStreamGenerator::generate(connection, crypt).await?;
+    Ok(ClientQuicConn::new(connection))
+}
 
-    let r = async {
-        match read_stream.process(main_sender_tx).await {
-            Ok(_) => info!("Tcp tunnel core task read stream finished"),
-            Err(e) => error!("Tcp tunnel core task read stream error: {:?}", e),
-        }
-    };
-
-    let w = async {
-        match write_stream.process(msg_stream).await {
-            Ok(_) => info!("Tcp tunnel core task write stream finished"),
-            Err(e) => error!("Tcp tunnel core task write stream error: {:?}", e),
-        }
-    };
-
-    // join!(r, w);
-    tokio::select! {
-        _ = r => {
-            info!("Tcp tunnel core task read stream end");
-        }
-        _ = w => {
-            info!("Tcp tunnel core task write stream end");
-        }
-    };
-
-    info!("Tcp tunnel core task finished");
+async fn quic_tunnel_core_task<S>(
+    server_addr: String,
+    crypt: Arc<Box<dyn DataCrypt>>,
+    msg_stream: &mut S,
+    main_sender_tx: Sender<ClientMsg>,
+) -> Result<(), VpnError>
+where
+    S: Stream<Item = ClientMsg> + Unpin + Send + Sync + 'static,
+{
+    trace!("Tcp tunnel core task start");
+    let connection = get_quic_stream(server_addr).await?;
+    run_client(connection, crypt, msg_stream, main_sender_tx).await?;
+    info!("tunnel core task finished");
 
     Ok(())
 }